+++ docs/assets/theme/mkdocs/js/theme.js
@@ -1,3 +1,33 @@
+function manipulateHtml() {
+  var $headlines = $('.rst-content').find('blockquote');
+
+  $headlines.each(function() {
+    var $this = $(this);
+    var answer = $this.nextUntil('hr');
+    var $wrapper = $('<div class="answer">');
+
+    $wrapper.append(answer);
+    $this.addClass('question');
+    $this.after($wrapper);
+  });
+
+  toggleSlideState($headlines);
+}
+
+function toggleSlideState($headlines) {
+  var $answers = $('.answer');
+
+  $answers.slideUp();
+  $headlines.removeClass('question--active');
+
+  $headlines.on('click', function() {
+    var $this = $(this);
+
+    $this.next().slideToggle();
+    $this.toggleClass('question--active');
+  });
+}
+
 $( document ).ready(function() {
 
     // Shift nav in mobile when clicking the menu.
@@ -22,6 +52,9 @@ $( document ).ready(function() {
     hljs.initHighlightingOnLoad();
 
     $('table').addClass('docutils');
+
+    // === FAQ
+    manipulateHtml();
 });
 
 window.SphinxRtdTheme = (function (jquery) {