+++ src/Commands/Field/FieldListCommand.php
@@ -31,7 +31,9 @@ class FieldListCommand extends PwConnector
             ->setDescription('Lists all available fields.')
             ->addOption('all', null, InputOption::VALUE_NONE, 'Show built-in fields. By default, system/permanent fields are not shown.')
             ->addOption('template', null, InputOption::VALUE_REQUIRED, 'Filter by template. When selected, only the fields from a specific template will be shown.')
-            ->addOption('type', null, InputOption::VALUE_REQUIRED, 'Filter by field type. When specified, only fields of the selected type will be shown.');
+            ->addOption('tag', null, InputOption::VALUE_REQUIRED, 'Filter by tag. When selected, only the fields with a specific tag will be shown.')
+            ->addOption('type', null, InputOption::VALUE_REQUIRED, 'Filter by field type. When specified, only fields of the selected type will be shown.')
+            ->addOption('unused', null, InputOption::VALUE_NONE, 'Only list unused fields.');
     }
 
     /**
@@ -72,7 +74,9 @@ class FieldListCommand extends PwConnector
     private function getFilter($input) {
         $filter = array();
         $filter['all'] = $input->getOption('all') ? true : false;
+        $filter['unused'] = $input->getOption('unused') ? true : false;
         $filter['template'] = $input->getOption('template') ? $input->getOption('template') : '';
+        $filter['tag'] = $input->getOption('tag') ? $input->getOption('tag') : '';
         $filter['type'] = $input->getOption('type') ? PwTools::getProperFieldtypeName($input->getOption('type')) : '';
 
         return (object)$filter;
@@ -94,8 +98,14 @@ class FieldListCommand extends PwConnector
                 if ($field->name != 'title') continue;
             }
 
+            // filter unused fields
+            if ($filter->unused && $field->getTemplates()->count()) continue;
+
             // filter by template
             if ($filter->template && !$field->getTemplates()->has($filter->template)) continue;
+
+            // filter by tag
+            if ($filter->tag && !$this->fieldHasTag($field->tags, $filter->tag)) continue;
 
             // filter by field type
             if ($filter->type && $field->type != $filter->type) continue;
@@ -132,6 +142,23 @@ class FieldListCommand extends PwConnector
 
         ksort($content);
         return (object) array('count' => $count, 'content' => $content);
+    }
+
+    /**
+     * check whether field has specific tag
+     *
+     * @param string $fieldTags
+     * @param string $tagName
+     * @return boolean
+     */
+    protected function fieldHasTag($fieldTags, $tagName) {
+      $tags = explode(',', $fieldTags);
+
+      if ($tags && in_array($tagName, $tags)) {
+        $hasTag = true;
+      }
+
+      return isset($hasTag) ? $hasTag : false;
     }
 
 }