+++ src/Commands/Module/ModuleEnableCommand.php
@@ -46,7 +46,7 @@ class ModuleEnableCommand extends PwConnector
 
         foreach ($modules as $module) {
             $this->checkIfModuleExistsLocally($module, $output, $input);
-            if (wire('modules')->getModule($module, array('noPermissionCheck' => true))) {
+            if (wire('modules')->getModule($module, array('noPermissionCheck' => true, 'noInit' => true))) {
                 $output->writeln("<info>Module {$module} installed successfully.</info>");
             }
         }